Moved TenantConfiguration to multitenancy module
