+++ multitenancy/src/main/java/org/axonframework/extensions/multitenancy/configuration/TenantConfiguration.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.axonframework.extensions.multitenancy.autoconfig;
+package org.axonframework.extensions.multitenancy.configuration;
 
 /**
  * Enables static access to default {@code TENANT_CORRELATION_KEY} used to correlate tenant identifiers within {@link
